# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minions Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List user deals
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller, all]
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Deals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealsListResponse"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal (buy product)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/confirm:
    post:
      operationId: confirmDeal
      tags: [deals]
      summary: Confirm deal received
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
  /deals/{id}/dispute:
    post:
      operationId: disputeDeal
      tags: [deals]
      summary: Open dispute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeBody"
      responses:
        "200":
          description: Disputed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/review:
    post:
      operationId: leaveDealReview
      tags: [deals]
      summary: Leave review for deal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /admin/deals:
    get:
      operationId: adminListDeals
      tags: [admin]
      summary: List all deals (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealsListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        bio:
          type: string
        balance:
          type: number
        frozenBalance:
          type: number
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        totalVolume:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        sellerLevel:
          type: string
        telegramId:
          type: string
        refCode:
          type: string
        createdAt:
          type: integer
        lastActive:
          type: integer
      required: [id, username]
...
    Product:
      type: object
      properties:
        id:
          type: string
        sellerId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        categoryName:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        game:
          type: string
        server:
          type: string
        status:
          type: string
        views:
          type: integer
        soldCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        isFavorited:
          type: boolean
        isPromoted:
          type: boolean
        createdAt:
          type: integer
        updatedAt:
          type: integer
      required: [id, sellerId, title, price, category]
...
    Deal:
      type: object
      properties:
        id:
          type: string
        dealNumber:
          type: integer
        buyerId:
          type: string
        buyer:
          $ref: "#/components/schemas/User"
        sellerId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        amount:
          type: number
        sellerAmount:
          type: number
        commission:
          type: number
        status:
          type: string
          enum: [pending, paid, delivered, completed, disputed, cancelled, refunded]
        deliveryData:
          type: string
        buyerConfirmed:
          type: boolean
        autoCompleteAt:
          type: integer
        disputeReason:
          type: string
        createdAt:
          type: integer
      required: [id, buyerId, sellerId, productId, amount, status]
    CreateDealBody:
      type: object
      properties:
        productId:
          type: string
      required: [productId]
...
    DisputeBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    ReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]
    Review:
      type: object
      properties:
        id:
          type: string
        dealId:
          type: string
        reviewerId:
          type: string
        reviewer:
          $ref: "#/components/schemas/User"
        sellerId:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: integer
      required: [id, reviewerId, sellerId, rating]
    DealsListResponse:
      type: object
      properties:
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [deals, total]
...